#0: Consolidating the program factories
